fixed app not ready error
